Updated assignment
