+++ src/UI/AppLayout.jsx
@@ -1,8 +1,8 @@
 import styled from "styled-components";
 import Header from "../components/Header";
-import PreferencesForm from "../UI/PersonNewsFeedBar";
 import { Outlet } from "react-router-dom";
 import { useState } from "react";
+import NewsSourceModal from "./NewsSourceModal";
 
 const StyledAppLayout = styled.div`
   height: 100vh;
@@ -18,7 +18,7 @@ function AppLayout() {
   return (
     <StyledAppLayout>
       <Header toggleModal={toggleModal} />
-      <PreferencesForm isOpen={isModalOpen} toggleModal={toggleModal} />
+      <NewsSourceModal isOpen={isModalOpen} toggleModal={toggleModal} />
       <Outlet />
     </StyledAppLayout>
   );
